// src/app/interfaces/book.interface.ts
// Generated by https://quicktype.io

export interface Book {
  book: BookClass;
}

export interface BookClass {
  title: string;
  pages: number;
  genre: string;
  cover: string;
  synopsis: string;
  year: number;
  ISBN: string;
  author: Author;
}

export interface Author {
  name: string;
  otherBooks: string[];
}